backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==2.2.0




 sudo service postgresql stop